refactor: resolve imports types ans errors
