+++ packages/cli/src/utils/LoadConfigurationFile.ts
@@ -1,7 +1,7 @@
 import { promises as fsp } from 'node:fs';
 
 import { FileNotFoundError, UnknownError } from '#cli/errors';
-import type { ConfigurationFile } from '#cli/types/ConfigurationFile';
+import type { ConfigurationFile } from '#cli/types';
 import keysToCamel from './KeysToCamel';
 
 /**
